requirements.txt
firebase-admin
Flask
flask-cors
python-magic-bin
python-dotenv